requirements.txt
pillow == 10.0.1
numpy == 1.26.1
opencv-python == 4.8.1.78
python == 3.12.0
